requirements.txt
flask==2.3.3
langchain==0.1.0
langchain-google-genai==1.0.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23